+++ Avatar 3.0/Dependencies/VRChat/Editor/ControlPanel/TBASdkChangelog.cs
@@ -30,9 +30,9 @@ UnityEditor.EditorPrefs.GetFloat("SDKColor_A")
 );
 
         EditorGUILayout.BeginHorizontal(boxGuiStyle, GUILayout.Height(26));
-        if (GUILayout.Button("The Black Arms Discord"))
+        if (GUILayout.Button("The Lonely Souls Discord"))
         {
-            Application.OpenURL("https://discord.gg/m6UfHkY");
+            Application.OpenURL("https://discord.gg/ryDrr4m");
         }
         if (GUILayout.Button("TGE VRC Asset Server"))
         {
@@ -42,9 +42,9 @@ UnityEditor.EditorPrefs.GetFloat("SDKColor_A")
         {
             Application.OpenURL("https://discord.gg/hEV4yKZ");
         }
-        if (GUILayout.Button("TBA Guilded"))
+        if (GUILayout.Button("TBA Discord"))
         {
-            Application.OpenURL("https://www.guilded.gg/i/Kk57LVQE");
+            Application.OpenURL("https://discord.gg/WNYc7x7sBC");
         }
         EditorGUILayout.EndHorizontal();
         EditorGUILayout.BeginHorizontal(boxGuiStyle, GUILayout.Height(26));
@@ -86,6 +86,11 @@ My SDX, this decision is final and wont be changed.
 
 This Branch is ENTIRELY SEPERATED from the main branch
 Therefore this will have its own changelog
+
+The Black Arms Avatar 3.0 SDX 2.0.3
+TBA got banned again so I updated links
+SDX Support server is now SDX community server
+New Main Servers (Invite via TBA Discord & Lonely Souls Buttons)
 
 The Black Arms Avatar 3.0 SDX 2.0.2
 Integrated Sentinels Importer (No Errors)