+++ VRCSDK/Dependencies/VRChat/Scripts/TBA_Discord.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class TBA_Discord : MonoBehaviour
 {
 
-    private string url = "https://discord.gg/m6UfHkY";
+    private string url = "https://discord.gg/WNYc7x7sBC";
 
     public void OpenURL()
     {